requirements-dev.txt
-e .
coverage==4.0.3
django-coverage-plugin==1.3
flake8==2.5.4
isort==4.2.5
mccabe==0.4.0
pydocstyle==1.0.0
pep8==1.7.0
pep8-naming==0.3.3
pre-commit==0.8.0
py==1.4.31
pytest==2.9.1
